Smoothed staircase movement
We now interpolate between the block heights to generate a "smooth" movement across stairways. This might be the best we can do with the current "blocky" collision detection, but there are some artifacts, especially when moving along (instead of across) diagonal staircases.

I think in the long run, we need to "smooth out" the blocks already in the collision detection.