Replaced the game loop with an async/await capable dispatcher
(This code was salvaged from an earlier refactoring attempt I did a couple of months ago.)

OpenGL and all window event handling need to be run from a single main thread, but using async/await, we can offload expensive things to worker threads. For async/await to work, we only need to implement a custom SynchronizationContext which works with our game loop.

This allows us to render and react to user events while connecting to the server, loading data, etc.