Run the game directly in the editor process
When you press "Run" in the editor, we no longer launch the game client as a child process, but rather run the game directly in the same process. This required a lot of fixes to the startup/shutdown code, but allows us to conveniently debug the game while adjusting values in the editor. The editor is now the default startup project.