Replace the use for raw floats/doubles for elapsed times with stronger types
GameTime is a new struct that represents a time value, always measured with a high-precision timer.

When you subtract two GameTimes, you get a GameDuration. Most of the uses of doubles to represent elapsed seconds since the last frame were replaced with GameDurations. This increases the type safety of our method calls (you can't accidentally pass a double value that's not meant as a duration), and makes the code read more naturally in several places. It also reduces confusion about the unit in which time values are stored.

Since the combination of GameTime and GameDuration together replaces the functionality of the old PrecisionTimer class, PrecisionTimer was removed.