Cleaned up channel event handling
In "single player" (when launching the client from the editor), we now connect the client directly to the universe's channel, without going through TCP. This revealed some ugly hacks in the startup sequence (for example, the first spawn was done by the client), which required a bit of clean-up. In general, the distribution of work between client and server is still not very clear.