Added a ChunkPos type
Instead of having to use raw X, Y, Z values, there's now a type called "ChunkPos" which represents the coordinates of a chunk. This simplifies the code in several places. Using this value also allows us to replace the custom "Dictionary3D<T>" with the much faster "Dictionary<ChunkPos, T>".